fix: remove unnecessary conditional
