+++ src/app/drive/components/DriveExplorer/DriveExplorer.tsx
@@ -379,12 +379,6 @@ const DriveExplorer = (props: DriveExplorerProps): JSX.Element => {
     setFolderInputKey(Date.now());
   };
 
-  const onViewModeButtonClicked = (): void => {
-    const setViewMode: FileViewMode = viewMode === FileViewMode.List ? FileViewMode.Grid : FileViewMode.List;
-
-    dispatch(storageActions.setViewMode(setViewMode));
-  };
-
   const onCreateFolderButtonClicked = (): void => {
     errorService.addBreadcrumb({
       level: 'info',
@@ -406,26 +400,6 @@ const DriveExplorer = (props: DriveExplorerProps): JSX.Element => {
     }
   };
 
-  const viewModesIcons = {
-    [FileViewMode.List]: (
-      <SquaresFour
-        size={24}
-        className="outline-none"
-        data-tooltip-id="viewMode-tooltip"
-        data-tooltip-content={translate('drive.viewMode.gridMode')}
-        data-tooltip-place="bottom"
-      />
-    ),
-    [FileViewMode.Grid]: (
-      <Rows
-        size={24}
-        className="outline-none"
-        data-tooltip-id="viewMode-tooltip"
-        data-tooltip-content={translate('drive.viewMode.listMode')}
-        data-tooltip-place="bottom"
-      />
-    ),
-  };
   const viewModes = {
     [FileViewMode.List]: DriveExplorerList,
     [FileViewMode.Grid]: DriveExplorerGrid,
@@ -649,111 +623,109 @@ const DriveExplorer = (props: DriveExplorerProps): JSX.Element => {
 
             <div className="flex shrink-0 flex-row">
               {!isTrash && (
-                <>
-                  <div className="flex items-center justify-center">
-                    <Menu as="div" className={openedWithRightClick ? '' : 'relative'}>
-                      {({ open, close }) => {
-                        useEffect(() => {
-                          if (!open) {
-                            setOpenedWithRightClick(false);
-                            setPosX(0);
-                            setPosY(0);
-                          }
-                        }, [open]);
+                <div className="flex items-center justify-center">
+                  <Menu as="div" className={openedWithRightClick ? '' : 'relative'}>
+                    {({ open, close }) => {
+                      useEffect(() => {
+                        if (!open) {
+                          setOpenedWithRightClick(false);
+                          setPosX(0);
+                          setPosY(0);
+                        }
+                      }, [open]);
 
-                        return (
-                          <>
-                            <Menu.Button ref={menuButtonRef as LegacyRef<HTMLButtonElement>}>
-                              <Button variant="primary" className="hidden">
-                                <div className="flex items-center justify-center space-x-2.5">
-                                  <span className="font-medium">{translate('actions.upload.new')}</span>
-                                  <div className="flex items-center space-x-0.5">
-                                    <Plus weight="bold" className="h-4 w-4" />
-                                    <CaretDown weight="fill" className="h-3 w-3" />
-                                  </div>
+                      return (
+                        <>
+                          <Menu.Button ref={menuButtonRef as LegacyRef<HTMLButtonElement>}>
+                            <Button variant="primary" className="hidden">
+                              <div className="flex items-center justify-center space-x-2.5">
+                                <span className="font-medium">{translate('actions.upload.new')}</span>
+                                <div className="flex items-center space-x-0.5">
+                                  <Plus weight="bold" className="h-4 w-4" />
+                                  <CaretDown weight="fill" className="h-3 w-3" />
                                 </div>
-                              </Button>
-                            </Menu.Button>
-                            <Transition
-                              className="absolute"
-                              enter="transition origin-top-right duration-100 ease-out"
-                              enterFrom="scale-95 opacity-0"
-                              enterTo="scale-100 opacity-100"
-                              leave="transition origin-top-right duration-100 ease-out"
-                              leaveFrom="scale-95 opacity-100"
-                              leaveTo="scale-100 opacity-0"
-                              style={
-                                openedWithRightClick ? { top: posY, left: posX, zIndex: 99 } : { right: 0, zIndex: 99 }
-                              }
-                            >
-                              {open && (
-                                <Menu.Items
-                                  className={
-                                    'mt-1 rounded-md border border-black/8 bg-white py-1.5 text-base shadow-subtle-hard outline-none'
-                                  }
-                                >
-                                  <Menu.Item>
-                                    {({ active }) => {
-                                      useHotkeys('shift+F', () => {
-                                        if (open) {
-                                          close();
-                                          onCreateFolderButtonClicked();
-                                        }
-                                      });
+                              </div>
+                            </Button>
+                          </Menu.Button>
+                          <Transition
+                            className="absolute"
+                            enter="transition origin-top-right duration-100 ease-out"
+                            enterFrom="scale-95 opacity-0"
+                            enterTo="scale-100 opacity-100"
+                            leave="transition origin-top-right duration-100 ease-out"
+                            leaveFrom="scale-95 opacity-100"
+                            leaveTo="scale-100 opacity-0"
+                            style={
+                              openedWithRightClick ? { top: posY, left: posX, zIndex: 99 } : { right: 0, zIndex: 99 }
+                            }
+                          >
+                            {open && (
+                              <Menu.Items
+                                className={
+                                  'mt-1 rounded-md border border-black/8 bg-white py-1.5 text-base shadow-subtle-hard outline-none'
+                                }
+                              >
+                                <Menu.Item>
+                                  {({ active }) => {
+                                    useHotkeys('shift+F', () => {
+                                      if (open) {
+                                        close();
+                                        onCreateFolderButtonClicked();
+                                      }
+                                    });
 
-                                      return (
-                                        <div
-                                          onClick={onCreateFolderButtonClicked}
-                                          className={`${
-                                            active && 'bg-gray-5'
-                                          } flex cursor-pointer items-center space-x-3 whitespace-nowrap py-2 pl-3 pr-5 text-gray-80 hover:bg-gray-5`}
-                                        >
-                                          <FolderSimplePlus size={20} />
-                                          <p>{translate('actions.upload.folder')}</p>
-                                          <span className="ml-5 flex grow items-center justify-end text-sm text-gray-40">
-                                            <ArrowFatUp size={14} /> F
-                                          </span>
-                                        </div>
-                                      );
-                                    }}
-                                  </Menu.Item>
-                                  <div className="mx-3 my-px flex border-t border-gray-5" />
-                                  <Menu.Item>
-                                    {({ active }) => (
+                                    return (
                                       <div
-                                        onClick={onUploadFileButtonClicked}
+                                        onClick={onCreateFolderButtonClicked}
                                         className={`${
                                           active && 'bg-gray-5'
                                         } flex cursor-pointer items-center space-x-3 whitespace-nowrap py-2 pl-3 pr-5 text-gray-80 hover:bg-gray-5`}
                                       >
-                                        <FileArrowUp size={20} />
-                                        <p className="ml-3">{translate('actions.upload.uploadFiles')}</p>
+                                        <FolderSimplePlus size={20} />
+                                        <p>{translate('actions.upload.folder')}</p>
+                                        <span className="ml-5 flex grow items-center justify-end text-sm text-gray-40">
+                                          <ArrowFatUp size={14} /> F
+                                        </span>
                                       </div>
-                                    )}
-                                  </Menu.Item>
-                                  <Menu.Item>
-                                    {({ active }) => (
-                                      <div
-                                        onClick={onUploadFolderButtonClicked}
-                                        className={`${
-                                          active && 'bg-gray-5'
-                                        } flex cursor-pointer items-center space-x-3 whitespace-nowrap py-2 pl-3 pr-5 text-gray-80 hover:bg-gray-5`}
-                                      >
-                                        <UploadSimple size={20} />
-                                        <p className="ml-3">{translate('actions.upload.uploadFolder')}</p>
-                                      </div>
-                                    )}
-                                  </Menu.Item>
-                                </Menu.Items>
-                              )}
-                            </Transition>
-                          </>
-                        );
-                      }}
-                    </Menu>
-                  </div>
-                  {!isTrash && <DriveTopBarItems />}
-                </>
+                                    );
+                                  }}
+                                </Menu.Item>
+                                <div className="mx-3 my-px flex border-t border-gray-5" />
+                                <Menu.Item>
+                                  {({ active }) => (
+                                    <div
+                                      onClick={onUploadFileButtonClicked}
+                                      className={`${
+                                        active && 'bg-gray-5'
+                                      } flex cursor-pointer items-center space-x-3 whitespace-nowrap py-2 pl-3 pr-5 text-gray-80 hover:bg-gray-5`}
+                                    >
+                                      <FileArrowUp size={20} />
+                                      <p className="ml-3">{translate('actions.upload.uploadFiles')}</p>
+                                    </div>
+                                  )}
+                                </Menu.Item>
+                                <Menu.Item>
+                                  {({ active }) => (
+                                    <div
+                                      onClick={onUploadFolderButtonClicked}
+                                      className={`${
+                                        active && 'bg-gray-5'
+                                      } flex cursor-pointer items-center space-x-3 whitespace-nowrap py-2 pl-3 pr-5 text-gray-80 hover:bg-gray-5`}
+                                    >
+                                      <UploadSimple size={20} />
+                                      <p className="ml-3">{translate('actions.upload.uploadFolder')}</p>
+                                    </div>
+                                  )}
+                                </Menu.Item>
+                              </Menu.Items>
+                            )}
+                          </Transition>
+                        </>
+                      );
+                    }}
+                  </Menu>
+                  <DriveTopBarItems />
+                </div>
               )}
               <DriveTopBarActions
                 hasAnyItemSelected={hasAnyItemSelected}
@@ -765,17 +737,6 @@ const DriveExplorer = (props: DriveExplorerProps): JSX.Element => {
                 setEditNameItem={setEditNameItem}
                 hasItems={hasItems}
               />
-              {!isTrash && (
-                <>
-                  {separatorV}
-                  <div className="flex items-center justify-center">
-                    <Button variant="tertiary" className="aspect-square" onClick={onViewModeButtonClicked}>
-                      {viewModesIcons[viewMode]}
-                    </Button>
-                    <TooltipElement id="viewMode-tooltip" delayShow={DELAY_SHOW_MS} />
-                  </div>
-                </>
-              )}
             </div>
             {isTrash && (
               <div className="flex items-center justify-center">
